Defines schema for sql migrations
This ensures that all tables are working in the default/supplied schema.
Before some where being migrated into a default schema, and some into
logary. Depending on the database this may work, but it definitely fails
on sql server.

Also adds a up/down test that works on a localhost SQL Server instance
to ensure that it still works.
This is commented to not break existing infrastructure