Format errors according to Stackdriver recommendations
Stackdriver error reporting can automatically pick up errors from log
entries if they are formatted properly:

"When logging error data from App Engine, Kubernetes Engine or Compute
Engine, the only requirement is for the log entry to contain the full
error message and stack trace. It should be logged as a multi-line
textPayload or in the message field of jsonPayload"

"C#: Must be the return value of Exception.ToString()"

https://cloud.google.com/error-reporting/docs/formatting-error-messages

This commit changes the formatting of errors to conform with that.